requirements.txt
Flask==3.0.3
SQLAlchemy==2.0.32
python-dotenv==1.0.1
boto3==1.34.162
Pillow==10.4.0
gunicorn==22.0.0

